backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis Cache
redis==5.0.1
hiredis==2.3.2

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Rate Limiting
slowapi==0.1.9

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0
respx==0.20.2
faker==22.0.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Production Server
gunicorn==21.2.0

# Environment
python-dotenv==1.0.0
